+++ index.php
@@ -59,6 +59,7 @@
 				spawnProperty('localhost_debug', $localhost ? "true" : "false");
 				spawnProperty('forum_user_data', $_COOKIE[$GLOBALS["COOKIE_NAME_USER_DATA"]]);
 				spawnProperty('forum_user_sign', $_COOKIE[$GLOBALS["COOKIE_NAME_USER_SIGN"]]);
+				spawnProperty('forum_path', authPath());
 			?>
         </x-properties>